Move ProbEvidenceMarginalMAP to experimental package
It remains to be seen if we should move UniverseState to experimental as
well. This would be more challenging because UniverseState requires
access to several private[language] fields.
